+++ ibig-core/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `count_ones`: population count
+
 ## [0.0.1] - 2026-06-05
 
 Work in progress. Low-level big-integer algorithms operating on slices of digits.